Created a new frontend component for dynamic table visualizations
